New nodes : apply suggestions and split nodes into multiple files
